requirements.txt
fastapi
uvicorn
langfuse
openai
python-dotenv
